+++ cdk_framework/EKS/lib/resource_constructs/general_constructs/general-sample-app-deployment-construct.ts
@@ -3,7 +3,7 @@ import { Cluster, FargateCluster } from 'aws-cdk-lib/aws-eks';
 import { PushModeSampleAppDeploymentConstruct } from '../specific_constructs/push-mode-sample-app-construct';
 import { PullModeSampleAppDeploymentConstruct } from '../specific_constructs/pull-mode-sample-app-construct';
 import { NamespaceConstruct } from '../specific_constructs/namespace-construct';
-import { SampleAppServiceConstruct } from './sample-app-service-construct';
+import { SampleAppServiceConstruct } from '../specific_constructs/sample-app-service-construct';
 
 
 export class GeneralSampleAppDeploymentConstruct extends Construct {
